refactor(computer-use): align plugin identity
