+++ plugins/bundle/computer-use/computer_use/client.py
