+++ plugins/bundle/computer-use/computer_use/lifecycle.py
