+++ plugins/bundle/computer-use/computer_use/protocol.py
