+++ plugins/bundle/computer-use/computer_use/transport/base.py
