+++ plugins/bundle/computer-use/computer_use/transport/unix_socket.py
