+++ tests/unit/plugins/computer_use/test_desktop_contention.py
@@ -16,9 +16,9 @@ from typing import Any
 
 import pytest
 
-from computer_use_tool import client as client_module
-from computer_use_tool.client import ComputerUseClient
-from computer_use_tool.protocol import ComputerUseProtocolError
+from computer_use import client as client_module
+from computer_use.client import ComputerUseClient
+from computer_use.protocol import ComputerUseProtocolError
 
 
 class _Transport:
