+++ tests/unit/plugins/computer_use/test_turn_lifecycle.py
@@ -17,14 +17,14 @@ from typing import Any
 
 import pytest
 
-from computer_use_tool import client as client_module
-from computer_use_tool.client import (
+from computer_use import client as client_module
+from computer_use.client import (
     ComputerUseClient,
     end_computer_use_turn,
     get_computer_use_client,
 )
-from computer_use_tool.protocol import ComputerUseProtocolError
-from computer_use_tool.transport import (
+from computer_use.protocol import ComputerUseProtocolError
+from computer_use.transport import (
     ComputerUseTransport,
     ReverseRequestHandler,
 )
